+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- feat: Add Memoizable module
+
 ## [0.1.0] - 2021-12-22
 
 - Initial release